Refactored controllers and routes to different folders
