+++ server/controllers/authControllers.js
@@ -1,25 +1,13 @@
 import userModel from "../models/userModel.js";
-import About from "../models/aboutModel.js";
 import utils from "../utils/encrypt.js"
 import jwt from "jsonwebtoken"; 
-import Experience from "../models/experienceModel.js";
-import { s3Client, uploadFileToS3, deleteFileFromS3 } from "../utils/s3FileUpload.js";
-import { GetObjectCommand } from "@aws-sdk/client-s3";
-import { v4 as uuidv4 } from 'uuid'
-import Project from "../models/projectModel.js";
-import Contact from "../models/contactModel.js";
-import { getSignedUrl } from "@aws-sdk/s3-request-presigner"
 
 
 
 
 const test = (req, res) => { 
     res.json("test works");
-   
 }
-/******************************************************
- *               PORTAL LOGIN FUNCTIONS        *
- ******************************************************/
 
 const loginUser = async (req, res) => { 
     try { 
@@ -111,606 +99,6 @@ const getProfile = (req,res) => {
     }
 }
 
-
-
-
-/******************************************************
- *              ABOUT SECTION FUNCTIONS        *
- ******************************************************/
-
-
-const getBio = async (req, res) => {
-    try {
-        const bio = await About.findOne();
-        if (!bio) {
-            return res.status(404).json({ message: 'Bio not found' });
-        }
-        res.status(200).json(bio);
-    } catch (error) {
-        console.error('Error fetching bio:', error);
-        res.status(500).json({ message: 'Server error' });
-    }
-};
-
-// Controller to update the bio
-const updateBio = async (req, res) => {
-    const { about } = req.body;
-
-    if (!about) {
-        return res.status(400).json({ message: 'About field is required' });
-    }
-
-    try {
-        let bio = await About.findOne();
-
-        if (bio) {
-            bio.about = about;
-            bio = await bio.save();
-        } else {
-            bio = new About({ about });
-            bio = await bio.save();
-        }
-
-        res.status(200).json(bio);
-    } catch (error) {
-        console.error('Error updating bio:', error);
-        res.status(500).json({ message: 'Server error' });
-    }
-};
-
-
-const getAboutImage = async(req, res) => { 
-    const { key } = req.params;
-    
-    try {
-        const params = {
-            Bucket: process.env.S3_BUCKET_NAME,
-            Key: key,
-        };
-
-        const command = new GetObjectCommand(params);
-        const s3Response = await s3Client.send(command);
-
-        res.setHeader('Content-Type', s3Response.ContentType);
-        s3Response.Body.pipe(res);
-    } catch (error) {
-        console.error('Error fetching image from S3:', error.message);
-        res.status(500).json({ error: 'Failed to fetch image' });
-    }
-}
-
-
-const getAboutImages = async(req,res) => {
-    try {
-      
-        const about = await About.findOne();
-      if (!about) return res.status(404).json({ images: [null, null, null, null] });
-  
-      // Ensure the returned array always has 4 slots
-      const images = [null, null, null, null];
-      
-      about.images.forEach((key, index) => {
-        images[index] = key
-          ? `${req.protocol}://${req.get("host")}/about/image/${encodeURIComponent(key)}`
-          : null;
-      });
-  
-      res.status(200).json({ images });
-    } catch (error) {
-      console.error("Error fetching about images:", error.message);
-      res.status(500).json({ error: "Failed to fetch about images" });
-    }
-};
-
-
-
-
-const updateAboutImage = async (req, res) => {
-    try {
-        const { index } = req.body; // The index of the image slot (0 to 3)
-        console.log('Request body:', req.body);
-        console.log('Uploaded file:', req.file);
-
-        // Validate index and file
-        if (index === undefined || index < 0 || index > 3) {
-            return res.status(400).json({ error: "Valid image slot index (0-3) is required" });
-        }
-        if (!req.file) {
-            return res.status(400).json({ error: "Image file is required" });
-        }
-
-        // Upload the image to S3
-        const s3Response = await uploadFileToS3(req.file, 'about');
-        console.log('S3 Upload Response:', s3Response);
-        const newImageKey = s3Response.key; // S3 key for the new image
-
-        // Fetch the `About` document
-        let about = await About.findOne();
-
-        // Get the key of the previous image (if any) to delete
-        const previousImageKey = about.images[index];
-
-        // Update the specific slot with the new S3 key
-        about.images[index] = newImageKey;
-
-        // Save the updated `About` document
-        await about.save();
-
-        // Delete the previous image from S3 (if it existed)
-        if (previousImageKey) {
-            await deleteFileFromS3(previousImageKey);
-        }
-
-        res.status(200).json({ message: "Image updated successfully", key: newImageKey });
-    } catch (error) {
-        console.error("Error updating about image:", error.message || error);
-        res.status(500).json({ error: "Failed to update about image" });
-    }
-};
-
-
-
-
-/******************************************************
- *               EXPERIENCE SECTION FUNCTIONS        *
- ******************************************************/
-
-
-const deleteExperience = async (req, res) => {
-    try {
-        const { publicId } = req.params;
-
-        // Find and delete the experience by publicId
-        const deletedExperience = await Experience.findOneAndDelete({ publicId });
-
-        if (deletedExperience) {
-            const s3Key = deletedExperience.companyPicture; // Assuming companyPicture holds the S3 object key
-
-            if (s3Key) {
-                // Use the helper function to delete the image from S3
-                await deleteFileFromS3(s3Key);
-                console.log("Company picture deleted from S3.");
-            }
-
-            res.status(200).json({ message: "Experience and corresponding image deleted successfully" });
-        } else {
-            res.status(404).json({ error: "Experience not found" });
-        }
-    } catch (error) {
-        console.error(error);
-        res.status(500).json({ error: "Failed to delete experience or image" });
-    }
-};
-
-const addExperience = async (req, res) => {
-    try {
-        const { companyName, role, responsibilities, startDate, endDate } = req.body;
-        
-        console.log('role', role)
-        console.log('Request body:', req.body);
-        console.log('Uploaded file:', req.file);
-
-        // Validate required fields
-        if (!companyName || !role || !responsibilities || !startDate || !endDate) {
-            return res.status(400).json({ error: "All fields are required" });
-        }
-
-        // Initialize the variable to store the S3 key
-        let companyPictureKey = null;
-
-        // Upload the file to S3 and get the key
-        if (req.file) {
-            const s3Response = await uploadFileToS3(req.file, 'experience');
-            console.log('S3 Upload Response:', s3Response);
-            companyPictureKey = s3Response.key; // Save the key to MongoDB
-        }
-
-        // Create a new experience document
-        const newExperience = new Experience({
-            companyName: companyName, 
-            companyPicture: companyPictureKey, // Save the S3 key
-            role,
-            responsibilities,
-            dateRange: {
-                startDate: new Date(startDate),
-                endDate: new Date(endDate),
-            },
-            publicId: uuidv4(),
-        });
-
-        // Save the new experience to the database
-        await newExperience.save();
-        res.status(201).json(newExperience);
-    } catch (error) {
-        console.error("Error adding experience:", error.message || error);
-        res.status(500).json({ error: "Failed to add experience" });
-    }
-};
-
-
-
-const updateExperience = async (req, res) => {
-    try {
-      const { publicId } = req.params; // Assuming the experience ID is passed as a URL parameter
-      const { companyName, role, responsibilities, startDate, endDate } = req.body;
-  
-      console.log('Request body:', req.body);
-      console.log('Uploaded file:', req.file);
-  
-      // Validate required fields
-      if (!companyName || !role || !responsibilities || !startDate || !endDate) {
-        console.log("something wrong here");
-        return res.status(400).json({ error: "All fields are required" });
-      }
-  
-      // Find the experience by publicId
-      let experience = await Experience.findOne({ publicId });
-      if (!experience) {
-        return res.status(404).json({ error: "Experience not found" });
-      }
-  
-      // Initialize the variable for the S3 key
-      let companyPictureKey = experience.companyPicture; // Keep the existing key unless a new image is uploaded
-      console.log("Current company picture key:", companyPictureKey);
-  
-      // Check if a new file is uploaded
-      if (req.file) {
-        console.log('file uplaoded!!')
-        const newFile = req.file;
-  
-        // If there's an existing image, delete it from S3
-        if (companyPictureKey && companyPictureKey !== newFile.originalname) {
-          const deleteResponse = await deleteFileFromS3(companyPictureKey);
-          console.log('S3 Delete Response:', deleteResponse);
-        }
-  
-        // Upload the new image to S3
-        const s3Response = await uploadFileToS3(newFile, 'experience');
-        console.log('S3 Upload Response:', s3Response);
-        companyPictureKey = s3Response.key; // Update the S3 key to the new image
-      }
-  
-      // Now update the experience in the database
-      experience = await Experience.findOneAndUpdate(
-        { publicId }, // Filter
-        {
-          companyName,
-          role,
-          responsibilities,
-          dateRange: {
-            startDate: new Date(startDate),
-            endDate: new Date(endDate),
-          },
-          companyPicture: companyPictureKey,
-        },
-        { new: true } // Return the updated document
-      );
-  
-      res.status(200).json(experience);
-    } catch (error) {
-      console.error("Error updating experience:", error.message || error);
-      res.status(500).json({ error: "Failed to update experience" });
-    }
-  };
-  
-
-const getExperiences = async (req, res) => {
-    try {
-        const experiences = await Experience.find().sort({ 'dateRange.startDate': -1 }); // Sort by date in ascending order
-
-        // Map through experiences to generate correct image URLs
-        const experiencesWithImageUrls = experiences.map(exp => {
-            return {
-                ...exp.toObject(),
-                companyPicture: `${req.protocol}://${req.get('host')}/experience/image/${encodeURIComponent(exp.companyPicture)}`
-            };
-        });
-
-        res.status(200).json(experiencesWithImageUrls);
-    } catch (error) {
-        console.error("Error retrieving experiences:", error.message || error);
-        res.status(500).json({ error: "Failed to retrieve experiences" });
-    }
-};
-
-const getExperienceImage = async (req, res) => { 
-    const { key } = req.params;
-    
-    try {
-        const params = {
-            Bucket: process.env.S3_BUCKET_NAME,
-            Key: key,
-        };
-
-        const command = new GetObjectCommand(params);
-        const s3Response = await s3Client.send(command);
-
-        res.setHeader('Content-Type', s3Response.ContentType);
-        s3Response.Body.pipe(res);
-    } catch (error) {
-        console.error('Error fetching image from S3:', error.message);
-        res.status(500).json({ error: 'Failed to fetch image' });
-    }
-};
-
-
-
-
-/******************************************************
- *               PROJECT SECTION FUNCTIONS        *
- ******************************************************/
-
-
-
-const addProject = async (req, res) => { 
-    try {
-        console.log(req.body)
-        const { title, description, githubURL, techStack, category, createdAt } = req.body;
-        
-        console.log(title, description, githubURL, techStack, category, createdAt)
-
-        console.log('Uploaded file:', req.file);
-
-    
-        if (!title || !description || !githubURL || !techStack || !category || !createdAt  ) {
-            return res.status(400).json({ error: "All fields are required" });
-        }
-        
-        // Initialize the variable to store the S3 key
-        let projectPictureKey = null;
-
-        // Upload the file to S3 and get the key
-        if (req.file) {
-            const s3Response = await uploadFileToS3(req.file, 'project');
-            console.log('S3 Upload Response:', s3Response);
-            projectPictureKey = s3Response.key; // Save the key to MongoDB
-        }
-
-         // Create a new experience document
-         const newProject = new Project({
-            title: title,
-            description: description,
-            githubURL: githubURL,
-            techStack: techStack,
-            category: category, 
-            createdAt: createdAt ? new Date(createdAt) : new Date(), 
-            pinned: false,
-            projectImage: projectPictureKey,
-            publicId: uuidv4()
-         })
-
-         await newProject.save()
-         res.status(201).json(newProject);
-    }
-    
-        catch (error) { 
-            console.error("Error adding project:", error.message || error);
-            res.status(500).json({ error: "Failed to add project" });
-        }
-     
-
-}
-
-const updateProject = async (req,res) => { 
-    try { 
-
-        const { publicId } = req.params
-
-        const { title,  description, githubURL, techStack, category, createdAt, pinned } = req.body
-
-        console.log("Received update request for Project ID:", publicId);
-        console.log("Received request body:", req.body);
-        console.log("Received file:", req.file ? req.file.originalname : "No file uploaded");
-
-
-        if (!title, !description, !githubURL, !techStack, !category, !createdAt, pinned == undefined) { 
-            console.log("All Arguements not given")
-            return res.status(400).json({error: "All fields are required"})
-        }
-    
-        //Find project by pubID
-        let project = await Project.findOne({ publicId  })
-        if (!project) {
-            return res.status(400).json({error: "Project not found"})
-        }
-
-        let projectPictureKey = project.projectImage
-        console.log(projectPictureKey)
-            // Check if a new file is uploaded
-        if (req.file) {
-            console.log('file uplaoded!!')
-            const newFile = req.file;
-    
-            // If there's an existing image, delete it from S3
-            if (projectPictureKey && projectPictureKey !== newFile.originalname) {
-            console.log('deleting')
-            const deleteResponse = await deleteFileFromS3(projectPictureKey);
-            console.log('S3 Delete Response:', deleteResponse);
-            }
-    
-            // Upload the new image to S3
-            const s3Response = await uploadFileToS3(newFile, 'project');
-            console.log('S3 Upload Response:', s3Response);
-            projectPictureKey = s3Response.key; // Update the S3 key to the new image
-        }
-
-        project = await Project.findOneAndUpdate({publicId}, 
-            {
-                title, 
-                description,
-                githubURL,
-                techStack,
-                category, 
-                createdAt, 
-                pinned,
-                projectImage: projectPictureKey
-
-            },
-            { new: true}
-         )
-
-        res.status(200).json(project)
-
-    } catch (error) { 
-        console.error("Error updating project:", error.message || error);
-        res.status(500).json({ error: "Failed to update project" });
-    }
-
-}
-
-const deleteProject = async (req, res) => {
-    try { 
-        const { publicId } = req.params
-
-        const deletedProject = await Project.findOneAndDelete({publicId})
-
-        if (deletedProject) { 
-            const s3Key = deletedProject.projectImage
-
-            if (s3Key) { 
-                await deleteFileFromS3(s3Key)
-                console.log("Company picture deleted from S3.");
-            }
-            res.status(200).json({ message: "Project and corresponding image deleted successfully" });
-        }
-        else { 
-            res.status(404).json({ error: "Project not found" });
-        }
-
-    } catch (error) { 
-        console.error(error)
-        res.status(500).json({ error: "Failed to delete project or image" });
-    }
-    
-}
-
-const getProjectImage = async (req,res) => {
-    const { key } = req.params;
-    
-    try {
-        const params = {
-            Bucket: process.env.S3_BUCKET_NAME,
-            Key: key,
-        };
-
-        const command = new GetObjectCommand(params);
-        const s3Response = await s3Client.send(command);
-
-        res.setHeader('Content-Type', s3Response.ContentType);
-        s3Response.Body.pipe(res);
-    } catch (error) {
-        console.error('Error fetching image from S3:', error.message);
-        res.status(500).json({ error: 'Failed to fetch image' });
-    }
-
-}
-
-const getProjects = async (req,res) => { 
-    try {
-        const projects = await Project.find()
-
-        // Map through experiences to generate correct image URLs
-        const projectsWithImageUrls = projects.map(proj => {
-            
-            console.log(proj.projectImage)
-            return {
-                ...proj.toObject(),
-                projectImage: `${req.protocol}://${req.get('host')}/projects/image/${encodeURIComponent(proj.projectImage)}`
-            };
-
-            
-    });
-
-    res.status(200).json(projectsWithImageUrls);
-
-    }
-    
-    
-    catch (error) {
-    console.error("Error retrieving experiences:", error.message || error);
-    res.status(500).json({ error: "Failed to retrieve experiences" });
-    }
-}
-
-/******************************************************
- *               CONTACT SECTION FUNCTIONS        *
- ******************************************************/
-
-const updateContact = async (req, res) => {
-    try {
-        const updates = req.body;
-        let resumeKey = null;
-
-        console.log("Received updates:", updates);
-        console.log("Received file:", req.file ? req.file.originalname : "No file uploaded");
-
-        let contact = await Contact.findOne();
-        if (!contact) {
-            contact = new Contact({ ...updates, resume: resumeKey });
-            await contact.save();
-        }
-
-        // Preserve the existing resume if no new file is uploaded
-        resumeKey = contact.resume; 
-
-        // Handle Resume Upload
-        if (req.file) {
-            console.log("Uploading new resume...");
-            const newFile = req.file;
-
-            // Delete old resume if it exists
-            if (contact.resume) {
-                console.log("Deleting old resume...");
-                await deleteFileFromS3(contact.resume);
-            }
-
-            // Upload new resume to S3
-            const s3Response = await uploadFileToS3(newFile, "contact");
-            console.log("S3 Upload Response:", s3Response);
-            resumeKey = s3Response.key; // Store S3 object key
-        }
-
-        // Ensure the update includes the existing resume
-        const updatedContact = await Contact.findOneAndUpdate(
-            {},
-            { 
-                $set: { ...updates, resume: resumeKey }  // Always include resume
-            },
-            { new: true, runValidators: true }
-        );
-
-        res.status(200).json({ success: true, updatedContact });
-    } catch (error) {
-        console.error("Error updating contact info:", error);
-        res.status(500).json({ error: "Failed to update contact info" });
-    }
-};
-
-
-const getContact = async(req,res) => { 
-    try { 
-        const contact = await Contact.findOne()
-        if (!contact) { 
-            return res.status(404).json({ message: "Contact info not found" });
-        }
-        let resumeUrl = null;
-        if (contact.resume) {
-            const params = { Bucket: process.env.S3_BUCKET_NAME, Key: contact.resume };
-            const command = new GetObjectCommand(params);
-            resumeUrl = await getSignedUrl(s3Client, command, { expiresIn: 3600 }); // 1-hour expiration
-        }
-
-        res.status(200).json({ ...contact.toObject(), resumeUrl });
-    } catch (error) { 
-        console.log(error)
-        return res.status(500).json( { message: "Could not fetch contact"})
-    }
-}
-
-
-
-
-
     
 
 export default { 
@@ -719,22 +107,5 @@ export default {
     loginUser, 
     getProfile, 
     logoutUser, 
-    getBio, 
-    updateBio, 
-    addExperience, 
-    deleteExperience,  
-    updateExperience, 
-    getExperiences, 
-    getExperienceImage, 
-    getAboutImage,
-    getAboutImages,
-    updateAboutImage,
-    getProjects,
-    getProjectImage,
-    addProject,
-    updateProject,
-    deleteProject,
-    getContact,
-    updateContact,
 
 }
