+++ server/models/aboutModel.js
@@ -1,12 +1,12 @@
-    import mongoose from "mongoose";
+import mongoose from "mongoose";
 
-    const aboutSchema = new mongoose.Schema({
-        about: {type: String, required: true},
-        images: [{ type: String }] 
-    })
+const aboutSchema = new mongoose.Schema({
+    about: {type: String, required: true},
+    images: [{ type: String }] 
+})
 
-    const About = mongoose.model("About",aboutSchema)
+const About = mongoose.model("About",aboutSchema)
 
-    export default About
+export default About
