+++ server/routes/authRoutes.js
@@ -1,6 +1,6 @@
 import express from "express"; 
 import cors from "cors"; 
-import  controllerFunctions from "../controllers/authControllers.js"
+import  authControllers from "../controllers/authControllers.js"
 import authenticateToken from "../middleware/auth.js";
 
 
@@ -15,12 +15,12 @@ router.use(
     })
 )
 
-router.get("/", controllerFunctions.test)
-router.post("/register", controllerFunctions.registerUser)
-router.post("/login",  controllerFunctions.loginUser)
-router.post("/logout", controllerFunctions.logoutUser)
-router.get("/profile", controllerFunctions.getProfile)
-router.get("/portal", authenticateToken, controllerFunctions.test)
+router.get("/", authControllers.test)
+router.post("/register", authControllers.registerUser)
+router.post("/login",  authControllers.loginUser)
+router.post("/logout", authControllers.logoutUser)
+router.get("/profile", authControllers.getProfile)
+router.get("/portal", authenticateToken, authControllers.test)
 
 export default router
