+++ server/routes/contactRoutes.js
@@ -1,12 +1,11 @@
 import express from "express";
-
-import controllerFunctions from "../controllers/authControllers.js";
+import contactControllers from "../controllers/contactControllers.js";
 import upload from "../middleware/upload.js";
 
 const router = express.Router();
 
 //Contact Routes
-router.get('/contact', controllerFunctions.getContact)
-router.put('/update-contact', upload.single('resume'), controllerFunctions.updateContact)
+router.get('/contact', contactControllers.getContact)
+router.put('/update-contact', upload.single('resume'), contactControllers.updateContact)
 
 export default router
